pyproject.toml
[project]
name = "ihsg-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "mplfinance>=0.12.10b0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "tensorflow>=2.20.0",
    "yfinance>=1.1.0",
]
